[IMPROVED] Moved unit tests to own folder
