+++ tests/unit/credentials/camunda_cloud_credentials_test.py
