+++ tests/unit/credentials/oauth_credentials_test.py
