+++ tests/unit/grpc_internals/zeebe_adapter_test.py
@@ -7,14 +7,14 @@ import grpc
 import pytest
 
 from pyzeebe.common.exceptions import *
-from pyzeebe.common.gateway_mock import GatewayMock
-from pyzeebe.common.random_utils import RANDOM_RANGE, random_task_context
 from pyzeebe.credentials.camunda_cloud_credentials import CamundaCloudCredentials
 from pyzeebe.credentials.oauth_credentials import OAuthCredentials
 from pyzeebe.grpc_internals.zeebe_adapter import ZeebeAdapter
 from pyzeebe.grpc_internals.zeebe_pb2 import *
 from pyzeebe.task.task import Task
 from pyzeebe.task.task_context import TaskContext
+from tests.unit.utils.gateway_mock import GatewayMock
+from tests.unit.utils.random_utils import RANDOM_RANGE, random_task_context
 
 zeebe_adapter: ZeebeAdapter
